requirements.txt
greenlet==3.2.4
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
torch>=2.2.0
transformers>=4.44.0

pandas>=2.0.0
scikit-learn>=1.3.0


google-generativeai
